Added permissions checks to API and added DELETE for folders API

See merge request exotic-matter/ftl-app!27

+++ ftl/core/views.py
@@ -1,16 +1,17 @@
 import json
 import os
 
-from django.contrib.auth.mixins import LoginRequiredMixin
+from django.contrib.auth.mixins import LoginRequiredMixin, PermissionRequiredMixin
 from django.http import HttpResponse
 from django.shortcuts import render, get_object_or_404
 from django.views import View
 from rest_framework import generics, views
 from rest_framework.authentication import SessionAuthentication
+from rest_framework.exceptions import ValidationError
 from rest_framework.parsers import MultiPartParser
 from rest_framework.response import Response
 
-from core.models import FTLDocument, FTLFolder
+from core.models import FTLDocument, FTLFolder, FTLModelPermissions
 from core.serializers import FTLDocumentSerializer, FTLFolderSerializer
 
 
@@ -24,8 +25,9 @@ class HomeView(LoginRequiredMixin, View):
         return render(request, 'core/home.html', context)
 
 
-class DownloadView(LoginRequiredMixin, View):
+class DownloadView(LoginRequiredMixin, PermissionRequiredMixin, View):
     http_method_names = ['get']
+    permission_required = ('core.view_ftldocument',)
 
     def get(self, request, *args, **kwargs):
         doc = get_object_or_404(FTLDocument.objects.filter(ftl_user=self.request.user, pid=kwargs['uuid']))
@@ -37,6 +39,7 @@ class DownloadView(LoginRequiredMixin, View):
 class FTLDocumentList(generics.ListAPIView):
     authentication_classes = (SessionAuthentication,)
     serializer_class = FTLDocumentSerializer
+    permission_classes = (FTLModelPermissions,)
 
     def get_queryset(self):
         current_folder = self.request.query_params.get('level', None)
@@ -55,6 +58,7 @@ class FTLDocumentDetail(generics.RetrieveUpdateDestroyAPIView):
     authentication_classes = (SessionAuthentication,)
     serializer_class = FTLDocumentSerializer
     lookup_field = 'pid'
+    permission_classes = (FTLModelPermissions,)
 
     def get_queryset(self):
         return FTLDocument.objects.filter(ftl_user=self.request.user)
@@ -66,20 +70,26 @@ class FTLDocumentDetail(generics.RetrieveUpdateDestroyAPIView):
         serializer.save(ftl_user=self.request.user)
 
     def perform_destroy(self, instance):
-        binary = instance.binary
-        super().perform_destroy(instance)
-        binary.file.close()
-        os.remove(binary.file.name)
+        if instance.org == self.request.user.org:
+            binary = instance.binary
+            super().perform_destroy(instance)
+            binary.file.close()
+            os.remove(binary.file.name)
+        else:
+            raise ValidationError("Trying to delete document from wrong user!")
 
 
 class FileUploadView(LoginRequiredMixin, views.APIView):
     authentication_classes = (SessionAuthentication,)
     parser_classes = (MultiPartParser,)
     serializer_class = FTLDocumentSerializer
+    permission_classes = (FTLModelPermissions,)
+    # Needed for applying permission checking on view that don't have any queryset
+    queryset = FTLDocument.objects.none()
 
     def post(self, request, *args, **kwargs):
         file_obj = request.data['file']
-        payload = json.loads(request.data['json'])  # Nothing for now
+        payload = json.loads(request.data['json'])
 
         # TODO check for empty form
 
@@ -104,6 +114,7 @@ class FTLFolderList(generics.ListCreateAPIView):
     authentication_classes = (SessionAuthentication,)
     serializer_class = FTLFolderSerializer
     pagination_class = None
+    permission_classes = (FTLModelPermissions,)
 
     def get_queryset(self):
         current_folder = self.request.query_params.get('level')
@@ -118,3 +129,22 @@ class FTLFolderList(generics.ListCreateAPIView):
 
     def perform_create(self, serializer):
         serializer.save(org=self.request.user.org)
+
+
+class FTLFolderDetail(generics.RetrieveUpdateDestroyAPIView):
+    authentication_classes = (SessionAuthentication,)
+    serializer_class = FTLFolderSerializer
+    lookup_field = 'id'
+    permission_classes = (FTLModelPermissions,)
+
+    def get_queryset(self):
+        return FTLFolder.objects.filter(org=self.request.user.org)
+
+    def perform_update(self, serializer):
+        serializer.save(org=self.request.user.org)
+
+    def perform_destroy(self, instance):
+        if instance.org == self.request.user.org:
+            instance.delete()
+        else:
+            raise ValidationError('Trying to delete a folder from wrong user!')
